sim: tighten modal-cards fast path (1.35% from 6%)
Initial bench against main showed BenchmarkAnnealRoundOnViseraiV4 regressed
+6.0% / p=0.001 / n=10 on hands with no ModalCards — well outside acceptable
overhead for the fast path.

Two fixes:

  1. Fast-path early exit in tryPitchOrdering. tupleCount is permutation-
     invariant (multiplication commutes), so it's lifted out of the closure
     and computed once per leaf. When tupleCount==1, the closure skips the
     mixed-radix mode decode entirely — pcBuf[i].Mode is already 0 from the
     per-permutation reset.

  2. Pack Mode (CardState) and modes (attackerMeta) as int8 instead of int.
     Both ride into the existing bool-block padding, so neither struct grows.
     pcBuf and the global cardMetaCache table are touched on every chain step;
     an extra cache line through either shows up at the bench level.

After: +1.35% sec/op, -0.07% B/op, allocs unchanged (n=10, p=0.001 sec/op).

The residual 1.35% is closure-pointer-call overhead — tryPitchOrdering still
dispatches through a function pointer rather than inlining. Recovering it
would require duplicating the Heap+pitch loops into a no-modal variant
(~80 lines), which isn't worth the maintenance cost vs the residual delta.

+++ internal/sim/cardmeta.go
@@ -37,9 +37,11 @@ type attackerMeta struct {
 	// Point — Instants and Attack Reactions (both 0 AP per FaB rules). Action cards and
 	// weapon swings cost 1 AP and don't set this.
 	isFreeChainStep bool
-	// modes is the mode count for a ModalCard, 1 for non-modal cards. Cached so the per-
-	// permutation mode-tuple loop avoids an interface assertion.
-	modes int
+	// modes is the mode count for a ModalCard, 1 for non-modal cards. Sized int8 so it
+	// packs into the bool block's padding without growing attackerMeta — every chain step
+	// reads permMeta[i] in the inner loop, and every extra cache line through that table
+	// shows up in the anneal bench.
+	modes int8
 }
 
 // costAt returns the card's effective cost given the current TurnState. Static cards return the
@@ -103,7 +105,7 @@ func cardMetaSlowPath(c Card, id ids.CardID) attackerMeta {
 		if n < 2 {
 			panic(fmt.Sprintf("ModalCard %s: Modes() = %d, want >= 2", c.Name(), n))
 		}
-		m.modes = n
+		m.modes = int8(n)
 	}
 	if vc, ok := c.(VariableCost); ok {
 		m.minCost = vc.MinCost()
